requirements.txt
psycopg2-binary~=2.9.3
kafka-python~=2.0.2
requests~=2.27.1
tabulate~=0.8.9

